+++ vrpc/VrpcBrowserAdapter.js
@@ -92,7 +92,7 @@ class VrpcAdapter {
     })
   }
 
- /**
+  /**
    * Registers an existing function and makes it callable from remote
    *
    * @param {Function} functionObject Existing function to be registered
@@ -399,7 +399,12 @@ class VrpcAdapter {
           calls.push({ id, val: v, err: e })
         }
       }
-      VrpcAdapter._handlePromise(json, Promise.all(calls))
+      this._handlePromise(
+        json,
+        Promise.allSettled(calls).then(result =>
+          result.filter(x => x.status === 'fulfilled').map(x => x.value)
+        )
+      )
     } catch (err) {
       json.e = err.message
     }
@@ -530,9 +535,10 @@ class VrpcAdapter {
     Object.entries(VrpcAdapter._listeners).forEach(([ik, iv]) => {
       Object.entries(iv).forEach(([ek, ev]) => {
         const { clients, listener, event } = ev
-        if (clients.has(clientId)) {
-          clients.delete(clientId)
-          if (clients.size === 0) {
+        const index = clients.indexOf(clientId)
+        if (index !== -1) {
+          clients.splice(index, 1)
+          if (clients.length === 0) {
             const instance = VrpcAdapter.getInstance(ik)
             if (instance && instance.removeListener) {
               instance.removeListener(event, listener)
@@ -584,11 +590,7 @@ class VrpcAdapter {
               eventId: arg,
               event: args[0]
             })
-            if (
-              VrpcAdapter.getInstance(context)
-                .listeners(arg)
-                .includes(listener)
-            ) {
+            if (!listener) {
               return null // skip call as listener is already registered
             }
             unwrapped.push(listener)
@@ -644,7 +646,7 @@ class VrpcAdapter {
       VrpcAdapter._listeners[instanceId] = {
         [eventId]: {
           event,
-          clients: new Set([clientId]),
+          clients: [clientId],
           listener: VrpcAdapter._generateListener({
             eventId,
             instanceId,
@@ -652,20 +654,22 @@ class VrpcAdapter {
           })
         }
       }
-    } else if (!VrpcAdapter._listeners[instanceId][eventId]) {
+      return VrpcAdapter._listeners[instanceId][eventId].listener
+    }
+    if (!VrpcAdapter._listeners[instanceId][eventId]) {
       VrpcAdapter._listeners[instanceId][eventId] = {
         event,
-        clients: new Set([clientId]),
+        clients: [clientId],
         listener: VrpcAdapter._generateListener({
           eventId,
           instanceId,
           isCallAll
         })
       }
-    } else {
-      VrpcAdapter._listeners[instanceId][eventId].clients.add(clientId)
+      return VrpcAdapter._listeners[instanceId][eventId].listener
     }
-    return VrpcAdapter._listeners[instanceId][eventId].listener
+    VrpcAdapter._listeners[instanceId][eventId].clients.push(clientId)
+    return null
   }
 
   static _generateListener ({ eventId, instanceId, isCallAll }) {
@@ -681,8 +685,9 @@ class VrpcAdapter {
       VrpcAdapter._listeners[instanceId][eventId]
     ) {
       const { listener, clients } = VrpcAdapter._listeners[instanceId][eventId]
-      clients.delete(clientId)
-      if (clients.size === 0) {
+      const index = clients.indexOf(clientId)
+      if (index !== -1) clients.splice(index, 1)
+      if (clients.length === 0) {
         delete VrpcAdapter._listeners[instanceId][eventId]
         if (Object.keys(VrpcAdapter._listeners[instanceId]).length === 0) {
           delete VrpcAdapter._listeners[instanceId]
@@ -696,9 +701,9 @@ class VrpcAdapter {
     const eventIds = VrpcAdapter._listeners[instanceId]
     if (!eventIds) return
     Object.entries(eventIds).forEach(([ek, ev]) => {
-      if (ev.event === event && ev.clients.has(clientId)) {
-        ev.clients.delete(clientId)
-        if (ev.clients.size === 0) {
+      if (ev.event === event && ev.clients.includes(clientId)) {
+        ev.clients = ev.clients.filter(x => x !== clientId)
+        if (ev.clients.length === 0) {
           VrpcAdapter.getInstance(instanceId).removeListener(
             ev.event,
             ev.listener
@@ -723,10 +728,7 @@ class VrpcAdapter {
   }
 
   static _generateId (object) {
-    return crypto
-      .createHash('md5')
-      .update(JSON.stringify(object))
-      .digest('hex')
+    return crypto.createHash('md5').update(JSON.stringify(object)).digest('hex')
   }
 
   static _getClassEntry (className) {